Fix signed/unsigned warnings in tests
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@25113 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/crypto/builtin/aes/aes-gen.c
@@ -21,7 +21,7 @@ aes_ctx ctx, dctx;
 
 static void init ()
 {
-    int i, j, r;
+    unsigned int i, j, r;
 
     srand(42);
     for (i = 0; i < 16; i++)
@@ -40,7 +40,7 @@ static void init ()
 
 static void hexdump(const unsigned char *ptr, size_t len)
 {
-    int i;
+    size_t i;
     for (i = 0; i < len; i++)
 	printf ("%s%02X", (i % 16 == 0) ? "\n    " : " ", ptr[i]);
 }
@@ -88,7 +88,7 @@ static void fips_test ()
 static void
 xor (unsigned char *out, const unsigned char *a, const unsigned char *b)
 {
-    int i;
+    unsigned int i;
     for (i = 0; i < B; i++)
 	out[i] = a[i] ^ b[i];
 }
@@ -96,7 +96,7 @@ xor (unsigned char *out, const unsigned char *a, const unsigned char *b)
 static void
 ecb_enc (unsigned char *out, unsigned char *in, unsigned int len)
 {
-    int i, r;
+    unsigned int i, r;
     for (i = 0; i < len; i += 16) {
 	r = aes_enc_blk (in + i, out + i, &ctx);
 	if (!r) fprintf(stderr, "error, line %d\n", __LINE__), exit(1);
@@ -107,7 +107,7 @@ ecb_enc (unsigned char *out, unsigned char *in, unsigned int len)
 static void
 ecb_dec (unsigned char *out, unsigned char *in, unsigned int len)
 {
-    int i, r;
+    unsigned int i, r;
     for (i = 0; i < len; i += 16) {
 	r = aes_dec_blk (in + i, out + i, &dctx);
 	if (!r) fprintf(stderr, "error, line %d\n", __LINE__), exit(1);
@@ -124,7 +124,7 @@ static void
 cbc_enc (unsigned char *out, unsigned char *in, unsigned char *iv,
 	 unsigned int len)
 {
-    int i, r;
+    unsigned int i, r;
     unsigned char tmp[B];
     D(iv);
     memcpy (tmp, iv, B);
@@ -144,7 +144,7 @@ static void
 cbc_dec (unsigned char *out, unsigned char *in, unsigned char *iv,
 	 unsigned int len)
 {
-    int i, r;
+    unsigned int i, r;
     unsigned char tmp[B];
     memcpy (tmp, iv, B);
     for (i = 0; i < len; i += B) {
@@ -230,7 +230,7 @@ cts_dec (unsigned char *out, unsigned char *in, unsigned char *iv,
 
 static void ecb_test ()
 {
-    int testno;
+    unsigned int testno;
     unsigned char tmp[4*B];
 
     printf ("ECB tests:\n");
@@ -261,7 +261,7 @@ unsigned char ivec[16] = { 0 };
 
 static void cbc_test ()
 {
-    int testno;
+    unsigned int testno;
     unsigned char tmp[4*B];
 
     printf ("CBC tests:\n");
@@ -290,7 +290,7 @@ static void cbc_test ()
 
 static void cts_test ()
 {
-    int testno;
+    unsigned int testno;
     unsigned char tmp[4*B];
 
     printf ("CTS tests:\n");
